day2/m5_m6_m7_m8/dtu_mlops_jan21/requirements.txt
-e .
click==8.0.3
matplotlib==3.5.1
numpy==1.21.0
pandas==1.3.5
python-dotenv==0.19.2
scikit_learn==1.0.2
setuptools==60.2.0
torch==1.10.0
hydra-core==1.1.1
google-cloud-storage==2.0.0